requirements.txt
authy==2.2.3
phonenumbers==8.10.16
python-dotenv==0.10.3
appdirs==1.4.3
asn1crypto==0.24.0
authy==2.2.3
bcrypt==3.1.7
blinker==1.4
braintree==3.53.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloud-init==19.4
command-not-found==0.3
configobj==5.0.6
cryptography==2.1.4
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.4.1
Django==2.2.10
django-allauth==0.39.1
django-audit-log==0.7.0
django-auditlog==0.4.7
django-crispy-forms==1.7.2
django-fernet-fields==0.6
django-hash-field==0.1
django-jsonfield==1.4.0
django-sslserver==0.22
Flask==1.1.1
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
gunicorn==19.6.0
httplib2==0.10.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==20.3
phonenumbers==8.10.16
Pillow==6.2.0
prettytable==0.7.2
psycopg2-binary==2.8.5
pyasn1==0.4.2
pycparser==2.20
pycryptodome==3.9.7
pycurl==7.43.0.1
pygobject==3.26.1
PyJWT==1.5.3
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.0
python-debian==0.1.32
python-dotenv==0.10.3
python-engineio==3.11.2
python-socketio==4.4.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==3.12
pyzbar==0.1.8
registry==0.4.2
requests==2.23.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
simplejson==3.10.0
six==1.14.0
sqlparse==0.3.0
ssh-import-id==5.7
stripe==2.44.0
systemd-python==234
tinyec==0.3.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
virtualenv==15.1.0
Werkzeug==0.16.0
whitenoise==3.2